# src/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pod Quiz Documentation",
    "description": "API documentation for Pod Project",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.2"
  },
  "servers": [
    {
      "url": "http://localhost:9200"
    },
    {
      "url": "http://172.20.10.12:9200"
    }
  ],
  "paths": {
...
    "/api/menuItem": {
      "post": {
        "tags": ["Manage-MenuItem"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Create a new menu item",
        "operationId": "createMenuItem",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Burger"
                  },
                  "price": {
                    "type": "number",
                    "example": 12.99
                  },
                  "description": {
                    "type": "string",
                    "example": "A juicy beef burger"
                  },
                  "ingredients": {
                    "type": "string",
                    "example": "Beef, Lettuce, Tomato, Bun"
                  },
                  "restaurantId": {
                    "type": "string",
                    "example": "60d0fe4f5311236168a109cb"
                  }
                },
                "required": [
                  "name",
                  "price",
                  "description",
                  "ingredients",
                  "restaurantId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Menu item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MenuItem"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/menuItem/{restaurantId}": {
      "get": {
        "tags": ["Manage-MenuItem"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get all menu items for a restaurant",
        "operationId": "getMenuItems",
        "parameters": [
          {
            "name": "restaurantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cb"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of menu items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MenuItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/menuItem/{id}": {
      "get": {
        "tags": ["Manage-MenuItem"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get a menu item by ID",
        "operationId": "getMenuItemById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Menu item details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MenuItem"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Menu item not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Menu item not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Manage-MenuItem"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Update a menu item by ID",
        "operationId": "updateMenuItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cc"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Burger"
                  },
                  "price": {
                    "type": "number",
                    "example": 12.99
                  },
                  "description": {
                    "type": "string",
                    "example": "A juicy beef burger"
                  },
                  "ingredients": {
                    "type": "string",
                    "example": "Beef, Lettuce, Tomato, Bun"
                  }
                },
                "required": [
                  "name",
                  "price",
                  "description",
                  "ingredients"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Menu item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MenuItem"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Menu item not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Menu item not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Manage-MenuItem"],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Delete a menu item by ID",
        "operationId": "deleteMenuItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d0fe4f5311236168a109cc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Menu item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Menu item deleted"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Menu item not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Menu item not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Server internal error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MenuItem": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Burger"
          },
          "price": {
            "type": "number",
            "example": 12.99
          },
          "description": {
            "type": "string",
            "example": "A juicy beef burger"
          },
          "ingredients": {
            "type": "string",
            "example": "Beef, Lettuce, Tomato, Bun"
          },
          "restaurant": {
            "type": "string",
            "example": "60d0fe4f5311236168a109cb"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "x-access-token",
        "in": "header"
      }
    }
  }
}